Improved YNAB duplicate handling
Calculate occurence for multiple transactions with the same amount
on the same day.  Occurrence is the third part of the "import_id" used
by YNAB for duplicate detection.

The calculation relies on having all transactions for a day.  To ensure
correctness, skip uploading all transactions from the first day in the list.