Rely on YNAB transactions' cleared state to match them
This changeset supports entering manual transactions in YNAB prior to
having the connector match up the bank statement. It consists of two
changes:

1- Retrieve bunq transactions as far back as the oldest uncleared
   transaction in YNAB; and,
2- Make sure to match any existing transaction that is still
   uncleared.

Note that we do not update existing transactions but rather submit
new transactions with a matching import-id to let YNAB handle the
actual matching. This approach gives a clue to the customer as to when
a manually entered transactions gets matched (YNAB shows the "matched icon"
and requires the match to be approved).

Fixes: #34