Use account notification
The API seems to have changed.  If you add an account notification, it will
add a user notifcation instead, and this will propagate to every account.

But when you remove a notification, it will only remove it from the account.
This will leave the user and other account notifications in place.

This commit switches to user notifcations, which when removed, do remove
all account notifcations.